loosen the tasks.md check for dependency_chain test; enforce task title format of tasks.md in skills.md; add anti-pattern for inputs id
