+++ tests/tasks/uipath-maestro-case/multi_stages_tasks/task_dependency_chain/task_dependency_chain.yaml
@@ -273,12 +273,20 @@ success_criteria:
     pass_threshold: 1.0
 
   - type: run_command
-    description: "Process exit=required-tasks-completed; Finalize exit=selected-tasks-completed; Finalize entry=selected-stage-completed; Done entry=selected-stage-completed on Finalize; tasks.md exposes activation-mode: sequential and entry-rule: runs-sequentially for First Step and Second Step; tasks carry runs-sequentially/current-stage-entered/selected-tasks-completed/adhoc task-entry rules; First Step and Second Step are consecutive single-task runs-sequentially task sets in Process data.tasks; First Step timer is non-repeating timeDuration (PT2S, no repeat); Second Step timer is non-repeating timeDuration (PT5S, no repeat); Region Check fires on current-stage-entered; root has string variable 'region' AND number variable 'priorityScore'; 'Optional Audit' is an adhoc-only process-typed skeleton task in Finalize"
+    description: "Process exit=required-tasks-completed; Finalize exit=selected-tasks-completed; Finalize entry=selected-stage-completed; Done entry=selected-stage-completed on Finalize; tasks carry runs-sequentially/current-stage-entered/selected-tasks-completed/adhoc task-entry rules; First Step and Second Step are consecutive single-task runs-sequentially task sets in Process data.tasks; First Step timer is non-repeating timeDuration (PT2S, no repeat); Second Step timer is non-repeating timeDuration (PT5S, no repeat); Region Check fires on current-stage-entered; root has string variable 'region' AND number variable 'priorityScore'; 'Optional Audit' is an adhoc-only process-typed skeleton task in Finalize"
     command: "python3 $TASK_DIR/check_task_dependency_chain.py"
     timeout: 720
     expected_exit_code: 0
     weight: 5.0
     pass_threshold: 1.0
+
+  - type: run_command
+    description: "(advisory, non-gating) tasks.md T-entries for First Step/Second Step quote the task display name and expose activation-mode: sequential / entry-rule: runs-sequentially per uipath-maestro-case SKILL.md Rule 6"
+    command: "python3 $TASK_DIR/check_task_dependency_chain_tasksmd.py"
+    timeout: 30
+    expected_exit_code: 0
+    weight: 1.0
+    pass_threshold: 0
 
 post_run:
   - command: "python3 $SKILLS_REPO_PATH/tests/tasks/uipath-maestro-case/_shared/cleanup_solutions.py"
